more tests, remove methods, and flatten structure
